pyproject.toml

[project]
name = "behave-automation"
version = "0.1.0"
description = "Automação BDD com Behave, Selenium e utilitários."
authors = [
    {name = "Ney Backes", email = "backesney@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "behavex>=4.6.0",
    "behave>=1.2.6,<2.0.0",
    "selenium>=4.15.2,<5.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "faker>=24.8.0,<25.0.0",
    "colorama>=0.4.6,<1.0.0"
]

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "allure-behave (>=2.0.0,<3.0.0)"

]   


[tool.taskipy.tasks]
run = 'behave'
lint = 'ruff check .'
format = 'ruff format .'
allure = 'behavex --formatter=behavex.outputs.formatters.allure_behavex_formatter:AllureBehaveXFormatter'
parallel = 'behavex --parallel-processes=3 --show-progress-bar'
pre_test = 'task lint'
test = 'behavex --parallel-processes=3 --show-progress-bar'

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'
